Hailstones Sequence (disclaimer: buggy code)
I found the famous Hailstones Sequence online and this made me curious of how I would build it from scratch. I tried and struggled (but I have not given up haha) and I'm caught in an endless loop and the program is not altering the initial number. So, this is my work so far :)